+++ transport/internal/safelycallhandler.go
@@ -18,7 +18,41 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package keyvalue
+package internal
 
-//go:generate thriftrw-go --plugin=yarpc kv.thrift
-//go:generate ../../../scripts/updateLicenses.sh
+import (
+	"fmt"
+	"log"
+	"runtime/debug"
+	"time"
+
+	"github.com/yarpc/yarpc-go/internal/errors"
+	"github.com/yarpc/yarpc-go/transport"
+
+	"golang.org/x/net/context"
+)
+
+// SafelyCallHandler calls the handler h, recovering panics and timeout errors,
+// converting them to yarpc errors. All other errors are passed trough.
+func SafelyCallHandler(h transport.Handler, start time.Time, ctx context.Context,
+	opts transport.Options, req *transport.Request, resq transport.ResponseWriter) (err error) {
+
+	// We recover panics from now on.
+	defer func() {
+		if r := recover(); r != nil {
+			log.Printf("Handler panicked: %v\n%s", r, debug.Stack())
+			err = fmt.Errorf("panic: %v", r)
+		}
+	}()
+
+	err = h.Handle(ctx, opts, req, resq)
+
+	// The handler stopped work on context deadline.
+	if err == context.DeadlineExceeded && err == ctx.Err() {
+		deadline, _ := ctx.Deadline()
+		err = errors.HandlerTimeoutError(req.Caller, req.Service,
+			req.Procedure, deadline.Sub(start))
+	}
+
+	return err
+}