+++ transport/tchannel/handler.go
@@ -28,6 +28,7 @@ import (
 	"github.com/yarpc/yarpc-go/internal/errors"
 	"github.com/yarpc/yarpc-go/internal/request"
 	"github.com/yarpc/yarpc-go/transport"
+	"github.com/yarpc/yarpc-go/transport/internal"
 
 	"github.com/uber/tchannel-go"
 	"golang.org/x/net/context"
@@ -76,6 +77,7 @@ func (c tchannelCall) Response() inboundCallResponse {
 type handler struct {
 	existing map[string]tchannel.Handler
 	Handler  transport.Handler
+	deps     transport.Deps
 }
 
 func (h handler) Handle(ctx context.Context, call *tchannel.InboundCall) {
@@ -89,7 +91,7 @@ func (h handler) Handle(ctx context.Context, call *tchannel.InboundCall) {
 
 func (h handler) handle(ctx context.Context, call inboundCall) {
 	start := time.Now()
-	err := h.callHandler(ctx, call)
+	err := h.callHandler(ctx, call, start)
 	if err == nil {
 		return
 	}
@@ -97,14 +99,6 @@ func (h handler) handle(ctx context.Context, call inboundCall) {
 	if _, ok := err.(tchannel.SystemError); ok {
 		call.Response().SendSystemError(err)
 		return
-	}
-
-	// The handler is well behaved and stopped work on context deadline. We
-	// forward this information to the client diligently.
-	if err == context.DeadlineExceeded && err == ctx.Err() {
-		deadline, _ := ctx.Deadline()
-		err = errors.HandlerTimeoutError(call.CallerName(), call.ServiceName(),
-			call.MethodString(), deadline.Sub(start))
 	}
 
 	err = errors.AsHandlerError(call.ServiceName(), call.MethodString(), err)
@@ -118,7 +112,7 @@ func (h handler) handle(ctx context.Context, call inboundCall) {
 	call.Response().SendSystemError(tchannel.NewSystemError(status, err.Error()))
 }
 
-func (h handler) callHandler(ctx context.Context, call inboundCall) error {
+func (h handler) callHandler(ctx context.Context, call inboundCall, start time.Time) error {
 	_, ok := ctx.Deadline()
 	if !ok {
 		return tchannel.ErrTimeoutRequired
@@ -137,6 +131,11 @@ func (h handler) callHandler(ctx context.Context, call inboundCall) error {
 	}
 	treq.Headers = headers
 
+	if tcall, ok := call.(tchannelCall); ok {
+		tracer := h.deps.Tracer()
+		ctx = tchannel.ExtractInboundSpan(ctx, tcall.InboundCall, headers.Items(), tracer)
+	}
+
 	body, err := call.Arg3Reader()
 	if err != nil {
 		return err
@@ -152,7 +151,7 @@ func (h handler) callHandler(ctx context.Context, call inboundCall) error {
 		return err
 	}
 
-	return h.Handler.Handle(ctx, transportOptions, treq, rw)
+	return internal.SafelyCallHandler(h.Handler, start, ctx, transportOptions, treq, rw)
 }
 
 type responseWriter struct {