+++ transport/tchannel/outbound.go
@@ -104,7 +104,12 @@ func (o outbound) Call(ctx context.Context, req *transport.Request) (*transport.
 	var err error
 
 	format := tchannel.Format(req.Encoding)
-	callOptions := tchannel.CallOptions{Format: format}
+	callOptions := tchannel.CallOptions{
+		Format:          format,
+		ShardKey:        req.ShardKey,
+		RoutingKey:      req.RoutingKey,
+		RoutingDelegate: req.RoutingDelegate,
+	}
 	if o.HostPort != "" {
 		// If the hostport is given, we use the BeginCall on the channel
 		// instead of the subchannel.
@@ -133,7 +138,10 @@ func (o outbound) Call(ctx context.Context, req *transport.Request) (*transport.
 		return nil, err
 	}
 
-	if err := writeRequestHeaders(ctx, format, req.Headers, call.Arg2Writer); err != nil {
+	// Inject tracing system baggage
+	reqHeaders := tchannel.InjectOutboundSpan(call.Response(), req.Headers.Items())
+
+	if err := writeRequestHeaders(ctx, format, reqHeaders, call.Arg2Writer); err != nil {
 		// TODO(abg): This will wrap IO errors while writing headers as encode
 		// errors. We should fix that.
 		return nil, encoding.RequestHeadersEncodeError(req, err)